refactor: extract agent and department mutation mixins from org_mutations.py
