+++ src/synthorg/api/services/org_mutations.py
@@ -16,37 +16,21 @@ from pydantic import BaseModel  # noqa: TC002
 
 from synthorg.api.concurrency import check_if_match, compute_etag
 from synthorg.api.dto_org import (  # noqa: TC001
-    CreateAgentOrgRequest,
-    CreateDepartmentRequest,
-    ReorderAgentsRequest,
-    ReorderDepartmentsRequest,
-    UpdateAgentOrgRequest,
     UpdateCompanyRequest,
     UpdateDepartmentRequest,
 )
 from synthorg.api.errors import (
     ApiValidationError,
-    ConflictError,
-    NotFoundError,
     VersionConflictError,
 )
-from synthorg.config.schema import AgentConfig
+from synthorg.api.services._org_agent_mutations import OrgAgentMutationsMixin
+from synthorg.api.services._org_department_mutations import OrgDepartmentMutationsMixin
+from synthorg.config.schema import AgentConfig  # noqa: TC001
 from synthorg.core.company import Company, Department
-from synthorg.core.enums import SeniorityLevel
 from synthorg.observability import get_logger
 from synthorg.observability.events.api import (
-    API_AGENT_CREATED,
-    API_AGENT_DELETED,
-    API_AGENT_UPDATED,
-    API_AGENTS_REORDERED,
     API_COMPANY_UPDATED,
     API_CONCURRENCY_CONFLICT,
-    API_DEPARTMENT_CREATED,
-    API_DEPARTMENT_DELETED,
-    API_DEPARTMENT_UPDATED,
-    API_DEPARTMENTS_REORDERED,
-    API_RESOURCE_CONFLICT,
-    API_RESOURCE_NOT_FOUND,
     API_VALIDATION_FAILED,
 )
 from synthorg.observability.events.versioning import VERSION_SNAPSHOT_FAILED
@@ -76,7 +60,7 @@ def _json_dump_models(models: Sequence[BaseModel]) -> str:
     )
 
 
-class OrgMutationService:
+class OrgMutationService(OrgAgentMutationsMixin, OrgDepartmentMutationsMixin):
     """Read-modify-write mutations on company/department/agent config.
 
     Args:
@@ -118,9 +102,6 @@ class OrgMutationService:
 
         Best-effort: versioning failures are logged but do not
         block the mutation.
-
-        Args:
-            saved_by: Actor who triggered the mutation.
         """
         if self._budget_versioning is None:
             return
@@ -139,14 +120,7 @@ class OrgMutationService:
             )
 
     async def _snapshot_company(self, saved_by: str) -> None:
-        """Snapshot the current Company structure if content changed.
-
-        Best-effort: versioning failures are logged but do not
-        block the mutation.
-
-        Args:
-            saved_by: Actor who triggered the mutation.
-        """
+        """Snapshot the current Company structure if content changed."""
         if self._company_versioning is None:
             return
         try:
@@ -175,13 +149,9 @@ class OrgMutationService:
         namespace: str,
         key: str,
     ) -> tuple[str, str]:
-        """Read a setting value and its ``updated_at`` for CAS.
-
-        Returns:
-            ``(value, updated_at)`` tuple, or ``("", "")`` if
-            the setting has no DB override.
-        """
-        return await self._settings.get_versioned(namespace, key)
+        """Read a setting value and its ``updated_at`` for CAS."""
+        result: tuple[str, str] = await self._settings.get_versioned(namespace, key)
+        return result
 
     async def _read_departments(self) -> tuple[Department, ...]:
         return await self._resolver.get_departments()
@@ -311,10 +281,7 @@ class OrgMutationService:
             return ""
 
     async def _company_snapshot_etag(self) -> str:
-        """Compute ETag for the full company snapshot.
-
-        Includes all writable company scalars plus agents and departments.
-        """
+        """Compute ETag for the full company snapshot."""
         name = await self._get_str_safe("company", "company_name")
         autonomy = await self._get_str_safe("company", "autonomy_level")
         budget = await self._get_str_safe("company", "total_monthly")
@@ -338,16 +305,9 @@ class OrgMutationService:
         if_match: str | None = None,
         saved_by: str = "api",
     ) -> tuple[dict[str, Any], str]:
-        """Update individual company scalar settings.
-
-        Args:
-            data: Partial update request.
-            if_match: If-Match header value for optimistic concurrency.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Returns:
-            Tuple of (updated fields dict, new ETag for full snapshot).
-        """
+        """Update individual company scalar settings."""
+        updated: dict[str, Any] = {}
+        new_etag = ""
         for attempt in range(_MAX_CAS_ATTEMPTS):
             try:
                 if if_match:
@@ -382,13 +342,7 @@ class OrgMutationService:
         self,
         data: UpdateCompanyRequest,
     ) -> dict[str, Any]:
-        """Atomically write all changed company scalars via set_many.
-
-        Preflights the version for every scalar that is about to
-        change, then persists them in a single transaction.  If any
-        CAS check fails the whole batch rolls back so ``update_company``
-        sees a consistent ``VersionConflictError`` and retries.
-        """
+        """Atomically write all changed company scalars via set_many."""
         items: list[tuple[str, str, str]] = []
         updated: dict[str, Any] = {}
         if data.company_name is not None:
@@ -419,715 +373,3 @@ class OrgMutationService:
             expected_updated_at_map=expected_map,
         )
         return updated
-
-    # ── Departments ───────────────────────────────────────────
-
-    async def create_department(
-        self,
-        data: CreateDepartmentRequest,
-        *,
-        saved_by: str = "api",
-    ) -> Department:
-        """Create a new department.
-
-        Args:
-            data: Department creation request.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Returns:
-            The created Department model.
-
-        Raises:
-            ConflictError: If a department with the same name exists.
-        """
-        for attempt in range(_MAX_CAS_ATTEMPTS):
-            try:
-                _, version = await self._read_setting_versioned(
-                    "company",
-                    "departments",
-                )
-                departments = await self._read_departments()
-                if self._find_department(departments, data.name):
-                    msg = f"Department {data.name!r} already exists"
-                    logger.warning(
-                        API_RESOURCE_CONFLICT,
-                        reason=msg,
-                        department=data.name,
-                    )
-                    raise ConflictError(msg)
-
-                dept = Department(
-                    name=data.name,
-                    head=data.head,
-                    budget_percent=data.budget_percent,
-                    autonomy_level=data.autonomy_level,
-                )
-                new_departments = (*departments, dept)
-                self._check_budget_sum(new_departments)
-                await self._write_departments(
-                    new_departments,
-                    expected_updated_at=version,
-                )
-                await self._snapshot_company(saved_by=saved_by)
-                break
-            except VersionConflictError:
-                if attempt == _MAX_CAS_ATTEMPTS - 1:
-                    logger.warning(
-                        API_CONCURRENCY_CONFLICT,
-                        resource="org_mutation",
-                        attempts=_MAX_CAS_ATTEMPTS,
-                    )
-                    raise
-                logger.debug(
-                    API_CONCURRENCY_CONFLICT,
-                    resource="org_mutation",
-                    attempt=attempt + 1,
-                    max_attempts=_MAX_CAS_ATTEMPTS,
-                )
-                continue
-
-        logger.info(
-            API_DEPARTMENT_CREATED,
-            department=data.name,
-            budget_percent=data.budget_percent,
-        )
-        return dept
-
-    async def update_department(
-        self,
-        name: str,
-        data: UpdateDepartmentRequest,
-        *,
-        if_match: str | None = None,
-        saved_by: str = "api",
-    ) -> Department:
-        """Update an existing department.
-
-        Args:
-            name: Current department name.
-            data: Partial update request.
-            if_match: If-Match header value for optimistic concurrency.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Returns:
-            The updated Department model.
-
-        Raises:
-            NotFoundError: If the department does not exist.
-        """
-        for attempt in range(_MAX_CAS_ATTEMPTS):
-            try:
-                _, version = await self._read_setting_versioned(
-                    "company",
-                    "departments",
-                )
-                departments = await self._read_departments()
-                existing = self._find_department(departments, name)
-                if existing is None:
-                    msg = f"Department {name!r} not found"
-                    logger.warning(
-                        API_RESOURCE_NOT_FOUND,
-                        reason=msg,
-                        department=name,
-                    )
-                    raise NotFoundError(msg)
-
-                if if_match:
-                    cur = json.dumps(
-                        existing.model_dump(mode="json"),
-                        sort_keys=True,
-                    )
-                    check_if_match(
-                        if_match,
-                        compute_etag(cur, ""),
-                        f"department:{name}",
-                    )
-
-                updates = self._collect_department_updates(data)
-                updated = existing.model_copy(update=updates, deep=True)
-                new_departments = tuple(
-                    updated if d.name.lower() == name.lower() else d
-                    for d in departments
-                )
-                self._check_budget_sum(new_departments)
-                await self._write_departments(
-                    new_departments,
-                    expected_updated_at=version,
-                )
-                await self._snapshot_company(saved_by=saved_by)
-                break
-            except VersionConflictError:
-                if attempt == _MAX_CAS_ATTEMPTS - 1:
-                    logger.warning(
-                        API_CONCURRENCY_CONFLICT,
-                        resource="org_mutation",
-                        attempts=_MAX_CAS_ATTEMPTS,
-                    )
-                    raise
-                logger.debug(
-                    API_CONCURRENCY_CONFLICT,
-                    resource="org_mutation",
-                    attempt=attempt + 1,
-                    max_attempts=_MAX_CAS_ATTEMPTS,
-                )
-                continue
-
-        logger.info(
-            API_DEPARTMENT_UPDATED,
-            department=name,
-            updated_fields=list(updates.keys()),
-        )
-        return updated
-
-    async def delete_department(
-        self,
-        name: str,
-        *,
-        saved_by: str = "api",
-    ) -> None:
-        """Delete a department.
-
-        Args:
-            name: Department name to delete.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Raises:
-            NotFoundError: If the department does not exist.
-            ConflictError: If the department has agents attached.
-        """
-        for attempt in range(_MAX_CAS_ATTEMPTS):
-            try:
-                await self._try_delete_department(name, saved_by=saved_by)
-                break
-            except VersionConflictError:
-                if attempt == _MAX_CAS_ATTEMPTS - 1:
-                    logger.warning(
-                        API_CONCURRENCY_CONFLICT,
-                        resource="org_mutation",
-                        attempts=_MAX_CAS_ATTEMPTS,
-                    )
-                    raise
-                logger.debug(
-                    API_CONCURRENCY_CONFLICT,
-                    resource="org_mutation",
-                    attempt=attempt + 1,
-                    max_attempts=_MAX_CAS_ATTEMPTS,
-                )
-                continue
-
-        logger.info(API_DEPARTMENT_DELETED, department=name)
-
-    async def _try_delete_department(
-        self,
-        name: str,
-        *,
-        saved_by: str,
-    ) -> None:
-        """One TOCTOU-safe attempt: CAS on BOTH departments and agents."""
-        _, dept_version = await self._read_setting_versioned("company", "departments")
-        _, agents_version = await self._read_setting_versioned("company", "agents")
-        departments = await self._read_departments()
-        if self._find_department(departments, name) is None:
-            msg = f"Department {name!r} not found"
-            logger.warning(API_RESOURCE_NOT_FOUND, reason=msg, department=name)
-            raise NotFoundError(msg)
-
-        agents = await self._read_agents()
-        self._check_no_attached_agents(name, agents)
-
-        new_departments = tuple(
-            d for d in departments if d.name.lower() != name.lower()
-        )
-        await self._settings.set_many(
-            [
-                ("company", "departments", _json_dump_models(new_departments)),
-                ("company", "agents", _json_dump_models(agents)),
-            ],
-            expected_updated_at_map={
-                ("company", "departments"): dept_version,
-                ("company", "agents"): agents_version,
-            },
-        )
-        await self._snapshot_company(saved_by=saved_by)
-
-    def _check_no_attached_agents(
-        self,
-        department_name: str,
-        agents: Sequence[AgentConfig],
-    ) -> None:
-        """Raise ConflictError when any agent references the department."""
-        attached = tuple(
-            a for a in agents if a.department.lower() == department_name.lower()
-        )
-        if not attached:
-            return
-        msg = (
-            f"Cannot delete department {department_name!r}: "
-            f"{len(attached)} agents attached"
-        )
-        logger.warning(
-            API_RESOURCE_CONFLICT,
-            reason=msg,
-            department=department_name,
-            agent_count=len(attached),
-        )
-        raise ConflictError(msg)
-
-    async def reorder_departments(
-        self,
-        data: ReorderDepartmentsRequest,
-        *,
-        saved_by: str = "api",
-    ) -> tuple[Department, ...]:
-        """Reorder departments.
-
-        Args:
-            data: Ordered list of department names.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Returns:
-            The reordered departments tuple.
-
-        Raises:
-            ApiValidationError: If names are not an exact permutation.
-        """
-        for attempt in range(_MAX_CAS_ATTEMPTS):
-            try:
-                _, version = await self._read_setting_versioned(
-                    "company",
-                    "departments",
-                )
-                departments = await self._read_departments()
-                current_names = tuple(d.name for d in departments)
-                self._validate_permutation(
-                    current_names,
-                    data.department_names,
-                    "department",
-                )
-
-                dept_by_lower = {d.name.lower(): d for d in departments}
-                reordered = tuple(
-                    dept_by_lower[n.lower()] for n in data.department_names
-                )
-                await self._write_departments(
-                    reordered,
-                    expected_updated_at=version,
-                )
-                await self._snapshot_company(saved_by=saved_by)
-                break
-            except VersionConflictError:
-                if attempt == _MAX_CAS_ATTEMPTS - 1:
-                    logger.warning(
-                        API_CONCURRENCY_CONFLICT,
-                        resource="org_mutation",
-                        attempts=_MAX_CAS_ATTEMPTS,
-                    )
-                    raise
-                logger.debug(
-                    API_CONCURRENCY_CONFLICT,
-                    resource="org_mutation",
-                    attempt=attempt + 1,
-                    max_attempts=_MAX_CAS_ATTEMPTS,
-                )
-                continue
-
-        logger.info(
-            API_DEPARTMENTS_REORDERED,
-            order=[d.name for d in reordered],
-        )
-        return reordered
-
-    # ── Agents ────────────────────────────────────────────────
-
-    async def create_agent(
-        self,
-        data: CreateAgentOrgRequest,
-        *,
-        saved_by: str = "api",
-    ) -> AgentConfig:
-        """Create a new agent in the org config.
-
-        Args:
-            data: Agent creation request.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Returns:
-            The created AgentConfig model.
-
-        Raises:
-            ApiValidationError: If the department does not exist.
-            ConflictError: If an agent with the same name exists.
-        """
-        for attempt in range(_MAX_CAS_ATTEMPTS):
-            try:
-                _, version = await self._read_setting_versioned(
-                    "company",
-                    "agents",
-                )
-                departments = await self._read_departments()
-                if not self._find_department(departments, data.department):
-                    msg = f"Department {data.department!r} does not exist"
-                    logger.warning(
-                        API_VALIDATION_FAILED,
-                        reason=msg,
-                        department=data.department,
-                    )
-                    raise ApiValidationError(msg)
-
-                agents = await self._read_agents()
-                if self._find_agent(agents, data.name):
-                    msg = f"Agent {data.name!r} already exists"
-                    logger.warning(
-                        API_RESOURCE_CONFLICT,
-                        reason=msg,
-                        agent=data.name,
-                    )
-                    raise ConflictError(msg)
-
-                model_dict: dict[str, Any] = {}
-                if data.model_provider is not None:
-                    model_dict = {
-                        "provider": str(data.model_provider),
-                        "model_id": str(data.model_id),
-                    }
-
-                agent = AgentConfig(
-                    name=data.name,
-                    role=data.role,
-                    department=data.department,
-                    level=data.level,
-                    model=model_dict,
-                )
-                new_agents = (*agents, agent)
-                await self._write_agents(
-                    new_agents,
-                    expected_updated_at=version,
-                )
-                await self._snapshot_company(saved_by=saved_by)
-                break
-            except VersionConflictError:
-                if attempt == _MAX_CAS_ATTEMPTS - 1:
-                    logger.warning(
-                        API_CONCURRENCY_CONFLICT,
-                        resource="org_mutation",
-                        attempts=_MAX_CAS_ATTEMPTS,
-                    )
-                    raise
-                logger.debug(
-                    API_CONCURRENCY_CONFLICT,
-                    resource="org_mutation",
-                    attempt=attempt + 1,
-                    max_attempts=_MAX_CAS_ATTEMPTS,
-                )
-                continue
-
-        logger.info(
-            API_AGENT_CREATED,
-            agent=data.name,
-            department=data.department,
-            level=data.level.value,
-        )
-        return agent
-
-    async def _validate_agent_update(
-        self,
-        name: str,
-        data: UpdateAgentOrgRequest,
-        agents: tuple[AgentConfig, ...],
-    ) -> dict[str, Any]:
-        """Validate agent update and collect field changes."""
-        updates: dict[str, Any] = {}
-        fields_set = data.model_fields_set
-
-        if "name" in fields_set and data.name is not None:
-            if self._find_agent(
-                tuple(a for a in agents if a.name.lower() != name.lower()),
-                str(data.name),
-            ):
-                msg = f"Agent {data.name!r} already exists"
-                logger.warning(
-                    API_RESOURCE_CONFLICT, reason=msg, agent_name=str(data.name)
-                )
-                raise ConflictError(msg)
-            updates["name"] = data.name
-
-        if "role" in fields_set and data.role is not None:
-            updates["role"] = data.role
-
-        if "department" in fields_set and data.department is not None:
-            departments = await self._read_departments()
-            if not self._find_department(departments, str(data.department)):
-                msg = f"Department {data.department!r} does not exist"
-                logger.warning(
-                    API_VALIDATION_FAILED, reason=msg, department=str(data.department)
-                )
-                raise ApiValidationError(msg)
-            updates["department"] = data.department
-
-        if "level" in fields_set and data.level is not None:
-            updates["level"] = data.level
-
-        if "autonomy_level" in fields_set:
-            updates["autonomy_level"] = data.autonomy_level
-
-        if "model_provider" in fields_set:
-            updates["model_provider"] = data.model_provider
-        if "model_id" in fields_set:
-            updates["model_id"] = data.model_id
-
-        return updates
-
-    async def update_agent(
-        self,
-        name: str,
-        data: UpdateAgentOrgRequest,
-        *,
-        if_match: str | None = None,
-        saved_by: str = "api",
-    ) -> AgentConfig:
-        """Update an existing agent.
-
-        Args:
-            name: Current agent name.
-            data: Partial update request.
-            if_match: If-Match header value for optimistic concurrency.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Returns:
-            The updated AgentConfig model.
-
-        Raises:
-            NotFoundError: If the agent does not exist.
-            ApiValidationError: If the target department does not exist.
-            ConflictError: If an agent with the new name already exists.
-        """
-        for attempt in range(_MAX_CAS_ATTEMPTS):
-            try:
-                _, version = await self._read_setting_versioned(
-                    "company",
-                    "agents",
-                )
-                agents = await self._read_agents()
-                existing = self._find_agent(agents, name)
-                if existing is None:
-                    msg = f"Agent {name!r} not found"
-                    logger.warning(
-                        API_RESOURCE_NOT_FOUND,
-                        reason=msg,
-                        agent=name,
-                    )
-                    raise NotFoundError(msg)
-
-                if if_match:
-                    cur = json.dumps(
-                        existing.model_dump(mode="json"),
-                        sort_keys=True,
-                    )
-                    check_if_match(
-                        if_match,
-                        compute_etag(cur, ""),
-                        f"agent:{name}",
-                    )
-
-                updates = await self._validate_agent_update(
-                    name,
-                    data,
-                    agents,
-                )
-
-                updated = existing.model_copy(update=updates, deep=True)
-                new_agents = tuple(
-                    updated if a.name.lower() == name.lower() else a for a in agents
-                )
-                await self._write_agents(
-                    new_agents,
-                    expected_updated_at=version,
-                )
-                await self._snapshot_company(saved_by=saved_by)
-                break
-            except VersionConflictError:
-                if attempt == _MAX_CAS_ATTEMPTS - 1:
-                    logger.warning(
-                        API_CONCURRENCY_CONFLICT,
-                        resource="org_mutation",
-                        attempts=_MAX_CAS_ATTEMPTS,
-                    )
-                    raise
-                logger.debug(
-                    API_CONCURRENCY_CONFLICT,
-                    resource="org_mutation",
-                    attempt=attempt + 1,
-                    max_attempts=_MAX_CAS_ATTEMPTS,
-                )
-                continue
-
-        logger.info(
-            API_AGENT_UPDATED,
-            agent=name,
-            updated_fields=list(updates.keys()),
-        )
-        return updated
-
-    async def delete_agent(self, name: str, *, saved_by: str = "api") -> None:
-        """Delete an agent from the org config.
-
-        Args:
-            name: Agent name to delete.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Raises:
-            NotFoundError: If the agent does not exist.
-            ConflictError: If the agent is the CEO (c_suite level).
-        """
-        for attempt in range(_MAX_CAS_ATTEMPTS):
-            try:
-                _, version = await self._read_setting_versioned(
-                    "company",
-                    "agents",
-                )
-                agents = await self._read_agents()
-                existing = self._find_agent(agents, name)
-                if existing is None:
-                    msg = f"Agent {name!r} not found"
-                    logger.warning(
-                        API_RESOURCE_NOT_FOUND,
-                        reason=msg,
-                        agent=name,
-                    )
-                    raise NotFoundError(msg)
-
-                if (
-                    existing.level == SeniorityLevel.C_SUITE
-                    and existing.role.lower() == "ceo"
-                ):
-                    msg = (
-                        f"Cannot delete CEO agent {name!r} -- reassign or demote first"
-                    )
-                    logger.warning(
-                        API_RESOURCE_CONFLICT,
-                        reason=msg,
-                        agent=name,
-                        level=existing.level.value,
-                        role=existing.role,
-                    )
-                    raise ConflictError(msg)
-
-                new_agents = tuple(a for a in agents if a.name.lower() != name.lower())
-                await self._write_agents(
-                    new_agents,
-                    expected_updated_at=version,
-                )
-                await self._snapshot_company(saved_by=saved_by)
-                break
-            except VersionConflictError:
-                if attempt == _MAX_CAS_ATTEMPTS - 1:
-                    logger.warning(
-                        API_CONCURRENCY_CONFLICT,
-                        resource="org_mutation",
-                        attempts=_MAX_CAS_ATTEMPTS,
-                    )
-                    raise
-                logger.debug(
-                    API_CONCURRENCY_CONFLICT,
-                    resource="org_mutation",
-                    attempt=attempt + 1,
-                    max_attempts=_MAX_CAS_ATTEMPTS,
-                )
-                continue
-
-        logger.info(API_AGENT_DELETED, agent=name)
-
-    async def reorder_agents(
-        self,
-        dept_name: str,
-        data: ReorderAgentsRequest,
-        *,
-        saved_by: str = "api",
-    ) -> tuple[AgentConfig, ...]:
-        """Reorder agents within a department.
-
-        Args:
-            dept_name: Department name.
-            data: Ordered list of agent names.
-            saved_by: Actor identity for version snapshot attribution.
-
-        Returns:
-            The reordered agents belonging to the department.
-
-        Raises:
-            NotFoundError: If the department does not exist.
-            ApiValidationError: If names are not an exact permutation.
-        """
-        for attempt in range(_MAX_CAS_ATTEMPTS):
-            try:
-                _, version = await self._read_setting_versioned(
-                    "company",
-                    "agents",
-                )
-                departments = await self._read_departments()
-                if not self._find_department(departments, dept_name):
-                    msg = f"Department {dept_name!r} not found"
-                    logger.warning(
-                        API_RESOURCE_NOT_FOUND,
-                        reason=msg,
-                        department=dept_name,
-                    )
-                    raise NotFoundError(msg)
-
-                agents = await self._read_agents()
-                dept_agents = tuple(
-                    a for a in agents if a.department.lower() == dept_name.lower()
-                )
-                current_names = tuple(a.name for a in dept_agents)
-                self._validate_permutation(
-                    current_names,
-                    data.agent_names,
-                    "agent",
-                )
-
-                agent_by_lower = {a.name.lower(): a for a in dept_agents}
-                reordered_dept = tuple(
-                    agent_by_lower[n.lower()] for n in data.agent_names
-                )
-
-                new_agents: list[AgentConfig] = []
-                dept_inserted = False
-                dept_lower = dept_name.lower()
-                for a in agents:
-                    if a.department.lower() == dept_lower:
-                        if not dept_inserted:
-                            new_agents.extend(reordered_dept)
-                            dept_inserted = True
-                    else:
-                        new_agents.append(a)
-                if not dept_inserted:
-                    new_agents.extend(reordered_dept)
-
-                await self._write_agents(
-                    tuple(new_agents),
-                    expected_updated_at=version,
-                )
-                await self._snapshot_company(saved_by=saved_by)
-                break
-            except VersionConflictError:
-                if attempt == _MAX_CAS_ATTEMPTS - 1:
-                    logger.warning(
-                        API_CONCURRENCY_CONFLICT,
-                        resource="org_mutation",
-                        attempts=_MAX_CAS_ATTEMPTS,
-                    )
-                    raise
-                logger.debug(
-                    API_CONCURRENCY_CONFLICT,
-                    resource="org_mutation",
-                    attempt=attempt + 1,
-                    max_attempts=_MAX_CAS_ATTEMPTS,
-                )
-                continue
-
-        logger.info(
-            API_AGENTS_REORDERED,
-            department=dept_name,
-            order=[a.name for a in reordered_dept],
-        )
-        return reordered_dept
